# Performance: Cache obj[key] access into 