Caching the value of `obj[key]` into a variable instead of continue access to object

PR
```
short string............................................. x 22,806,425 ops/sec ±0.47% (191 runs sampled)
unsafe short string...................................... x 1,027,602,256 ops/sec ±0.34% (191 runs sampled)
short string with double quote........................... x 13,276,009 ops/sec ±0.33% (190 runs sampled)
long string without double quotes........................ x 15,398 ops/sec ±0.24% (192 runs sampled)
unsafe long string without double quotes................. x 1,011,926,548 ops/sec ±0.30% (188 runs sampled)
long string.............................................. x 15,779 ops/sec ±0.22% (193 runs sampled)
unsafe long string....................................... x 1,018,123,202 ops/sec ±0.31% (190 runs sampled)
number................................................... x 1,021,065,239 ops/sec ±0.26% (191 runs sampled)
integer.................................................. x 218,888,807 ops/sec ±0.33% (191 runs sampled)
```
 master
```
short string............................................. x 23,010,004 ops/sec ±0.58% (190 runs sampled)
unsafe short string...................................... x 1,062,255,820 ops/sec ±0.37% (191 runs sampled)
short string with double quote........................... x 13,749,470 ops/sec ±0.37% (191 runs sampled)
long string without double quotes........................ x 15,661 ops/sec ±0.31% (192 runs sampled)
unsafe long string without double quotes................. x 1,062,491,836 ops/sec ±0.37% (191 runs sampled)
long string.............................................. x 16,092 ops/sec ±0.32% (192 runs sampled)
unsafe long string....................................... x 1,060,550,353 ops/sec ±0.31% (192 runs sampled)
number................................................... x 1,058,792,481 ops/sec ±0.33% (191 runs sampled)
integer.................................................. x 220,943,513 ops/sec ±0.29% (192 runs sampled)
```

#### Checklist

- [x] run `npm run test` and `npm run benchmark`
- [ ] tests and/or benchmarks are included
- [ ] documentation is changed or added
- [x] commit message and code follows the [Developer's Certification of Origin](https://github.com/fastify/.github/blob/master/CONTRIBUTING.md#developers-certificate-of-origin-11)
      and the [Code of conduct](https://github.com/fastify/.github/blob/master/CODE_OF_CONDUCT.md)