@nigrosimone 2024-04-10
The problem with micro-optimization is the instability of the benchmark. The PR is correct, but the difference is negligible. Tomorrow try to run benchmark on thousands of samples (now is 100)