@mcollina 2024-04-11
It seems you are cutting out all the object benchmarks? I thought this impacted those.