@cesco69 2024-04-11
> It seems you are cutting out all the object benchmarks? I thought this impacted those.
@mcollina **minSamples 500**

PR
```
short string............................................. x 23,486,212 ops/sec ±0.41% (591 runs sampled)
unsafe short string...................................... x 1,071,682,433 ops/sec ±0.25% (591 runs sampled)
short string with double quote........................... x 14,071,608 ops/sec ±0.31% (590 runs sampled)
long string without double quotes........................ x 15,706 ops/sec ±0.28% (590 runs sampled)
unsafe long string without double quotes................. x 1,078,410,601 ops/sec ±0.30% (592 runs sampled)
long string.............................................. x 16,323 ops/sec ±0.27% (592 runs sampled)
unsafe long string....................................... x 1,083,630,259 ops/sec ±0.31% (592 runs sampled)
number................................................... x 1,033,759,666 ops/sec ±0.27% (590 runs sampled)
integer.................................................. x 219,223,329 ops/sec ±0.27% (591 runs sampled)
formatted date-time...................................... x 1,697,386 ops/sec ±0.27% (590 runs sampled)
formatted date........................................... x 1,109,075 ops/sec ±0.27% (588 runs sampled)
formatted time........................................... x 1,107,261 ops/sec ±0.27% (590 runs sampled)
short array of numbers................................... x 77,795 ops/sec ±0.21% (591 runs sampled)
short array of integers.................................. x 65,291 ops/sec ±0.22% (591 runs sampled)
short array of short strings............................. x 20,105 ops/sec ±0.21% (590 runs sampled)
short array of long strings.............................. x 20,130 ops/sec ±0.19% (590 runs sampled)
short array of objects with properties of different types x 9,486 ops/sec ±0.19% (590 runs sampled)
object with number property.............................. x 1,044,304,818 ops/sec ±0.26% (589 runs sampled)
object with integer property............................. x 222,809,874 ops/sec ±0.27% (591 runs sampled)
object with short string property........................ x 22,944,390 ops/sec ±0.33% (589 runs sampled)
object with long string property......................... x 16,179 ops/sec ±0.25% (592 runs sampled)
object with properties of different types................ x 1,837,666 ops/sec ±0.42% (582 runs sampled)
simple object............................................ x 9,934,298 ops/sec ±0.24% (590 runs sampled)
simple object with required fields....................... x 9,819,406 ops/sec ±0.23% (590 runs sampled)
object with const string property........................ x 1,057,533,293 ops/sec ±0.29% (589 runs sampled)
object with const number property........................ x 1,055,692,892 ops/sec ±0.28% (588 runs sampled)
object with const bool property.......................... x 1,059,992,242 ops/sec ±0.29% (591 runs sampled)
object with const object property........................ x 1,077,156,612 ops/sec ±0.27% (591 runs sampled)
object with const null property.......................... x 1,080,980,241 ops/sec ±0.27% (591 runs sampled)
```
Master
```
short string............................................. x 20,336,896 ops/sec ±1.56% (591 runs sampled)
unsafe short string...................................... x 1,063,754,842 ops/sec ±0.27% (591 runs sampled)
short string with double quote........................... x 14,049,638 ops/sec ±0.29% (591 runs sampled)
long string without double quotes........................ x 15,826 ops/sec ±0.28% (591 runs sampled)
unsafe long string without double quotes................. x 1,078,118,074 ops/sec ±0.28% (591 runs sampled)
long string.............................................. x 16,363 ops/sec ±0.27% (592 runs sampled)
unsafe long string....................................... x 1,072,536,686 ops/sec ±0.25% (592 runs sampled)
number................................................... x 1,073,209,502 ops/sec ±0.29% (592 runs sampled)
integer.................................................. x 219,494,857 ops/sec ±0.32% (586 runs sampled)
formatted date-time...................................... x 1,626,708 ops/sec ±0.25% (588 runs sampled)
formatted date........................................... x 1,110,917 ops/sec ±0.27% (591 runs sampled)
formatted time........................................... x 1,113,999 ops/sec ±0.25% (589 runs sampled)
short array of numbers................................... x 77,544 ops/sec ±0.24% (591 runs sampled)
short array of integers.................................. x 64,974 ops/sec ±0.21% (591 runs sampled)
short array of short strings............................. x 20,208 ops/sec ±0.20% (591 runs sampled)
short array of long strings.............................. x 20,094 ops/sec ±0.26% (591 runs sampled)
short array of objects with properties of different types x 9,549 ops/sec ±0.16% (591 runs sampled)
object with number property.............................. x 1,043,371,336 ops/sec ±0.25% (590 runs sampled)
object with integer property............................. x 218,891,786 ops/sec ±0.18% (589 runs sampled)
object with short string property........................ x 23,101,644 ops/sec ±0.32% (589 runs sampled)
object with long string property......................... x 15,707 ops/sec ±0.11% (590 runs sampled)
object with properties of different types................ x 1,831,538 ops/sec ±0.52% (580 runs sampled)
simple object............................................ x 9,887,605 ops/sec ±0.23% (590 runs sampled)
simple object with required fields....................... x 10,118,325 ops/sec ±0.31% (590 runs sampled)
object with const string property........................ x 1,072,143,772 ops/sec ±0.28% (590 runs sampled)
object with const number property........................ x 1,061,290,681 ops/sec ±0.26% (590 runs sampled)
object with const bool property.......................... x 1,074,415,151 ops/sec ±0.29% (591 runs sampled)
object with const object property........................ x 1,080,103,573 ops/sec ±0.29% (591 runs sampled)
object with const null property.......................... x 1,074,310,277 ops/sec ±0.27% (589 runs sampled)
```